ASFIPSEC: Fixing miscellaneous error
In this patch following things are done.
1. GET_CACHE_ALLIGNED uint is changed to uintptr_t.
2. Removing compilation errors for IPSEC debug case.
3. Temp fix in case OutSA not found.
4. Fix for bug in asfctrl_xfrm_add_outsa, sa_id instead of outSA.ulSAContainerIndex.
5. pSelList = pSA->pHeadSelList; removes crash on adding AH policy.
6. SECFP_ENC_NONE solves the InvalidAuthEncAlgo problem.

CR:ENGR00331411
Reviewed-on: http://git.am.freescale.net:8181/18511
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>

+++ asfmodule/asfipsec/driver/ipsecfp_sec4x.c
@@ -42,6 +42,7 @@
 #ifdef ASF_SECFP_PROTO_OFFLOAD
 #include <pdb.h>
 #endif
+#include "dcl.h"
 extern struct device *pdev;
 
 #define xstr(s) str(s)
@@ -54,8 +55,8 @@ extern struct device *pdev;
 #define DESC_AEAD_GIVENCRYPT_TEXT_LEN 27
 #define OP_PCL_IPSEC_NULL 0x0b00
 
-#define GET_CACHE_ALLIGNED(x) (u32 *)(((int)x + (L1_CACHE_BYTES - 1)) \
-	& ~(L1_CACHE_BYTES - 1)) /* THIS CAN BE AN ISSUE FOR 64BIT PLATFORM */
+#define GET_CACHE_ALLIGNED(x) (uintptr_t *)(((uintptr_t)x + (L1_CACHE_BYTES - 1)) \
+	& ~(L1_CACHE_BYTES - 1))
 
 static void secfp_splitKeyDone(struct device *dev, u32 *desc, u32 error,
 				void *context)
